requirements.txt
# Core
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0

# ONNX Runtime (CPU default)
onnxruntime>=1.12.0

# Optional: faster assignment (Hungarian)
# lap>=0.4.0